backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiomysql
pydantic
pydantic-settings
python-dotenv
cryptography
aiosqlite
aiohttp
aiomysql
pymysql
